Move OSM node deduplication to an inner module.
